Add Embodied Emissions
This commit adds embodied emissions to the model. Add embodied emissions in the EmissionEmbodied table. These emissions are assumed to occur in the year of construction (i.e. vintage). These emissions are included in both the EmissionLimit constraint, and the objective function as they are used in the emission cost calculation.